requirements.txt
streamlit==1.31.0
langchain==0.0.335
langchain-core==0.1.16
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4
python-dotenv==1.0.0
openai==1.6.1
tiktoken==0.5.2
PyPDF2==3.0.0
pypdf==3.15.0
pydantic==1.10.13
pydantic-core==0.42.0
chromadb==0.4.13
docarray==0.30.0
huggingface-hub==0.19.0
sentence-transformers==2.2.0
